+++ tests/testthat/test-branch.length.R
@@ -1,20 +1,3 @@
-test_that("branch lengths are correctly calculated", {
-  library(ape)
-  set.seed(1)
-  tree <- rtree(n=5)
-
-  bls <- setNames(c(0.0617863, 1.0695542, 1.5364765, 1.2643343, 0.9235931 ),
-                  paste0("t", c(2,1,3,4,5)))
-
-  edge.info <- tip.root.path(tree)
-
-  expect_equal(round(species.branch.length(tree), 7),
-               bls)
-  expect_equal(round(species.branch.length(edge.info = edge.info), 7),
-               bls)
-
-})
-
 test_that("tip lengths are correctly calculated", {
   library(ape)
   set.seed(1)