+++ NEWS.md
@@ -2,10 +2,10 @@
 
 # phyloraster 2.1.0
 
-- added functions to compute full and tip branch lengths: 'species.tip.length()' and 'species.branch.length()'
+- added function to compute tip branch lengths: 'species.tip.length()'
 - added new function to compute ED for each species: 'species.ed()'
 - fixed ED calculation for raster data
-- added new tests for the new functions
+- added examples of the ED calculation on the vignette
 
 # phyloraster 2.0.1